Revert "hmi-controller: Convert from systemd to DLT logging"
This reverts commit 3f9a8d551f8845ab2bb7d42526dc876809c196ec.

When launching _any_ new app the HMI freezes up due to some
threading/locking issue in a DLT thread (possibly ivi_logging framework
also involved in the issue).

Reverting this, for some reason fixes the HMI symptom, so most apps
launch without freezing, but it seems that launching one of the apps
that also had their logging converted causes a similar crash.

This seemingly worked OK in previous GDP version so it might be a
complex interaction with some updated versions of the HMI related
frameworks.

This is a stopgap awaiting further investigation of the issue.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>